requirements.txt
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
Django==2.1
future==0.18.2
idna==2.8
line-bot-sdk==1.15.0
pytz==2019.3
requests==2.22.0
sqlparse==0.3.0
urllib3==1.25.7
gunicorn == 19.7.1
six
pymongo == 3.11.0
BeautifulSoup4 == 2.0.1